refactor(tests/integration): move to common tests package
